backend/requirements.txt
fastapi
uvicorn
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
python-multipart
supabase
pydantic-settings
